Test whether path leads to a directory

+++ file_utils.py
@@ -10,11 +10,16 @@ def get_file(path):
     except FileNotFoundError:
         contents = errors.FILE_NOT_EXIST
         error = True
+    except IsADirectoryError:
+        contents = errors.IS_A_DIRECTORY
+        error = True
     return contents, error
 
 def create_file(path, contents):
     if os.path.isfile(path):
         return errors.FILE_EXISTS, True
+    if os.path.isdir(path):
+        return errors.IS_A_DIRECTORY, True
     parent_directory = os.path.dirname(path)
     os.makedirs(parent_directory, exist_ok=True)
     with open(path, 'w') as f:
@@ -23,6 +28,8 @@ def create_file(path, contents):
     return contents, error
 
 def update_file(path, contents):
+    if os.path.isdir(path):
+        return errors.IS_A_DIRECTORY, True
     if not os.path.isfile(path):
         return errors.FILE_NOT_EXIST, True
     with open(path, 'w') as f:
@@ -31,6 +38,8 @@ def update_file(path, contents):
     return contents, error
 
 def delete_file(path):
+    if os.path.isdir(path):
+        return errors.IS_A_DIRECTORY, True
     try:
         os.remove(path)
         contents = ''
